+++ memory-viz/src/style.ts
@@ -124,6 +124,7 @@ function setStyleSheet(memory_model: MemoryModel, global_style?: string) {
         text.id { 
             fill: ${config.id_color};
             text-anchor: middle;
+            cursor: pointer;
         }
         text.type {
             fill: ${config.value_color};
@@ -135,6 +136,9 @@ function setStyleSheet(memory_model: MemoryModel, global_style?: string) {
         }
         path {
             stroke: ${config.rect_style?.stroke ?? "rgb(0,0,0)"};
+        }
+        .highlighted path {
+            fill: rgba(255, 255, 0, 0.6) !important;
         }
     `;