+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added unique object ids for SVG output object tags `<g>`
 - Added support for global CSS styling using the `--global-style` option in the CLI
 - Added optional attributes width and height for each drawn object
+- Added configurable hover interactivity to object ids
 
 ### 🐛 Bug fixes
 
@@ -41,7 +42,6 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Increased timeout for CLI tests from 2 seconds to 30 seconds
 - Added appropriate icons to the buttons on the demo and webstepper pages
 - Updated demo page tests in accordance with this change
-- Added configurable hover interactivity to object ids
 
 ### 🐛 Bug fixes